# api/api-spec.yaml
openapi: 3.0.3
info:
  title: TripleThink API
  description: |
    Event-sourced narrative construction system for multi-book fiction series.

    TripleThink enables authors to manage complex storylines with:
    - Epistemic tracking (who knows what, when)
    - Fiction management (false narratives with strict audience constraints)
    - Time-travel queries (entity state at any timestamp)
    - Token-efficient AI integration

    ## Key Concepts

    - **Entities**: Characters, events, objects, locations, systems
    - **Metadata**: Rich author notes and AI guidance (loaded on-demand)
    - **Knowledge States**: What characters believe at specific times
    - **Fictions**: False narratives with target audiences
    - **Epistemic Queries**: The power feature for tracking beliefs
  version: 1.0.0
  contact:
    name: TripleThink Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/narrative/book/{bookId}:
    get:
      tags: [Narrative]
      summary: Get book with structure
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book with acts and chapters
  /api/narrative/book/{bookId}/chapter/{chapterNum}:
    get:
      tags: [Narrative]
      summary: Get chapter by number
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: chapterNum
          in: path
          required: true
          schema:
            type: integer
        - name: include_scenes
          in: query
          schema:
            type: boolean
            default: true
        - name: include_events
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Chapter data
  /api/narrative/scene/{sceneId}:
    get:
      tags: [Narrative]
      summary: Get scene with epistemic context
      parameters:
        - name: sceneId
          in: path
          required: true
          schema:
            type: string
        - name: include_epistemic_context
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Scene with context
  /api/narrative/generate-outline:
    post:
      tags: [Narrative]
      summary: Generate chapter outline from events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [book_id]
              properties:
                book_id:
                  type: string
                chapter_range:
                  type: string
                  description: 'Format: start-end'
                  example: '1-5'
                character_focus:
                  type: string
      responses:
        '200':
          description: Generated outline

  # ==========================================
  # VALIDATION
  # ==========================================
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []